refactor(bayn): modularize Alpaca read adapter
